+++ be/src/util/jsonb_updater.h
@@ -224,7 +224,7 @@ private:
 
         if (value->isInt() && curr->isInt()) {
             // Both are ints and optimization can be done here
-            int64_t val = ((const JsonbIntVal*)value)->val();
+            int128_t val = ((const JsonbIntVal*)value)->val();
             // setVal may fail because the new value can't fit into the current one.
             if (((JsonbIntVal*)curr)->setVal(val)) {
                 return JsonbErrType::E_NONE;
